Initial test of rf_render_png in Python.
